feat(internal): tool rework, tool definition rework, allows tool to define use of stdout as output instead of mounted directory with file, stdout connection on engine still TODO

+++ internal/core/engine/engine.go
@@ -11,10 +11,10 @@ import (
 
 type ToolExecution struct {
 	Path string
-	Tool *tools.ToolDefinition
+	Tool *tools.Tool
 }
 
-func NewToolExecution(tool *tools.ToolDefinition, path string) (*ToolExecution, error) {
+func NewToolExecution(tool *tools.Tool, path string) (*ToolExecution, error) {
 	absPath, err := filepath.Abs(path)
 
 	if err != nil {
@@ -48,30 +48,46 @@ func (engine *InfrarunEngine) Execute(ctx context.Context, toolExecution *ToolEx
 		return nil, err
 	}
 
-	// Tempdir for output of container
-	outputDir, err := os.MkdirTemp("", "infrarun-")
-
-	if err != nil {
-		return nil, err
+	volumeBinds := []docker.VolumeBind{
+		{Host: toolExecution.Path, Guest: toolExecution.Tool.InputPath},
 	}
 
-	defer os.RemoveAll(outputDir)
+	outputDir := ""
 
-	err = engine.Backend.RunContainer(ctx, docker.ContainerInfo{
-		Image: toolExecution.Tool.Image,
-		Cmd:   toolExecution.Tool.Cmd,
-		VolumeBinds: []docker.VolumeBind{
-			{Host: outputDir, Guest: toolExecution.Tool.OutputPath},
-			{Host: toolExecution.Path, Guest: toolExecution.Tool.InputPath},
-		},
+	if !toolExecution.Tool.CaptureStdout {
+		// Tempdir for output of container
+		var err error
+		outputDir, err = os.MkdirTemp("", "infrarun-")
+
+		if err != nil {
+			return nil, err
+		}
+
+		defer os.RemoveAll(outputDir)
+
+		volumeBinds = append(volumeBinds, docker.VolumeBind{
+			Host:  outputDir,
+			Guest: toolExecution.Tool.OutputPath,
+		})
+	}
+
+	err := engine.Backend.RunContainer(ctx, docker.ContainerInfo{
+		Image:       toolExecution.Tool.Image,
+		Cmd:         toolExecution.Tool.Cmd,
+		VolumeBinds: volumeBinds,
 	})
 
 	if err != nil {
 		return nil, err
 	}
 
-	// TODO: allow post processing since some tools might not export to SARIF
-	outputFilePath := filepath.Clean(outputDir + "/" + toolExecution.Tool.OutputFile)
+	if toolExecution.Tool.CaptureStdout {
+		// TODO: implement this case
 
-	return os.ReadFile(outputFilePath)
+		return make([]byte, 0), nil
+	} else {
+		outputFilePath := filepath.Clean(outputDir + "/" + toolExecution.Tool.OutputFile)
+		return os.ReadFile(outputFilePath)
+	}
+
 }
